+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 - Fixed issue that caused invalid OBJ files to be produced for M2 collision geometry.
 - Added `skin.subMeshes.enabled` property to M2 metadata, indicating which submeshes were exported.
 - Added `groups.enabled` property to WMO metadata, indicating which WMO groups were exported.
+- Added `textures.fileNameInternal` property to M2 metadata, mapping fileDataID to a listfile entry.
+- Added `textures.fileNameExternal` property to M2 metadata, pointing to relative texture export location.
+- Added `textures.mtlName` property to M2 metadata, linking to an entry in the MTL file.
 
 0.1.34 (24-08-2021)
 - Added height textures to ADT alpha map exports, mapped in the metadata layer entries.